backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
werkzeug
google-generativeai
python-dotenv